package(compile-mdx): progress
