+++ apps/site/mdx.use.client.ts
@@ -1,5 +1,5 @@
-'use strict';
-// @augustin-note: this sould stay in the app, not in the mdx package
+import type { MDXComponents } from 'mdx/types';
+
 import Blockquote from './components/Common/Blockquote';
 import Button from './components/Common/Button';
 import LinkWithArrow from './components/Downloads/Release/LinkWithArrow';
@@ -10,8 +10,6 @@ import MDXImage from './components/MDX/Image';
 
 /**
  * A full list of React Components that we want to pass through to MDX
- *
- * @satisfies {import('mdx/types').MDXComponents}
  */
 export const clientMdxComponents = {
   // Renders MDX CodeTabs
@@ -20,12 +18,10 @@ export const clientMdxComponents = {
   Button: Button,
   // Links with External Arrow
   LinkWithArrow: LinkWithArrow,
-};
+} as MDXComponents;
 
 /**
  * A full list of wired HTML elements into custom React Components
- *
- * @type {import('mdx/types').MDXComponents}
  */
 export const htmlComponents = {
   // Renders a Link Component for `a` tags
@@ -36,4 +32,4 @@ export const htmlComponents = {
   pre: MDXCodeBox,
   // Renders an Image Component for `img` tags
   img: MDXImage,
-};
+} as MDXComponents;
