+++ apps/site/util/__tests__/gitHubUtils.test.mjs
@@ -2,6 +2,7 @@ import {
   getGitHubAvatarUrl,
   getGitHubBlobUrl,
   getGitHubApiDocsUrl,
+  createGitHubSlugger
 } from '@/util/gitHubUtils';
 
 describe('GitHub utils', () => {
@@ -23,5 +24,10 @@ describe('GitHub utils', () => {
     const expected =
       'https://api.github.com/repos/nodejs/node/contents/doc/api?ref=assert';
     expect(result).toBe(expected);
+  });
+
+  it('createGitHubSlugger returns a function', () => {
+    const slugger = createGitHubSlugger();
+    expect(typeof slugger).toBe('function');
   });
 });
