+++ apps/site/util/gitHubUtils.ts
@@ -1,3 +1,5 @@
+import GitHubSlugger from 'github-slugger';
+
 export const getGitHubAvatarUrl = (username: string): string =>
   `https://avatars.githubusercontent.com/${username}`;
 
@@ -6,3 +8,9 @@ export const getGitHubBlobUrl = (filename: string) =>
 
 export const getGitHubApiDocsUrl = (ref: string) =>
   `https://api.github.com/repos/nodejs/node/contents/doc/api?ref=${ref}`;
+
+export const createGitHubSlugger = () => {
+  const githubSlugger = new GitHubSlugger();
+
+  return (text: string) => githubSlugger.slug(text);
+};
