--- packages/compile-mdx/lib/utils/gitHubUtils.ts
@@ -1,7 +1,0 @@
-import GitHubSlugger from 'github-slugger';
-
-export const createGitHubSlugger = () => {
-  const githubSlugger = new GitHubSlugger();
-
-  return (text: string) => githubSlugger.slug(text);
-};
