+++ packages/compile-mdx/package.json
@@ -3,16 +3,14 @@
   "version": "1.0.0",
   "type": "module",
   "exports": {
-    ".": "./dist/index.js",
     "./compiler": "./dist/compiler.js",
     "./shiki": "./dist/shiki.js",
     "./utils/*": "./dist/utils/*"
   },
   "scripts": {
     "build": "tsc --declaration --declarationMap",
-    "type-check": "tsc --noEmit",
-    "lint:js": "eslint \"**/*.{js,mjs,ts}\"",
-    "test:unit": "node --import=tsx --test ./**/*.test.ts"
+    "check-types": "tsc --noEmit",
+    "lint:js": "eslint \"**/*.{js,mjs,ts}\""
   },
   "engines": {
     "node": ">=20"
@@ -21,7 +19,6 @@
     "eslint": "~9.10.0",
     "eslint-import-resolver-typescript": "~3.6.3",
     "eslint-plugin-import-x": "~4.2.1",
-    "tsx": "^4.19.1",
     "typescript": "~5.5.4",
     "typescript-eslint": "~8.5.0"
   },
